VST3ParametersWindow extracted into separate file
